pyproject.toml
[tool.poetry]
name = "exiftool"
version = "0.3.3"
description = "Role to install ExifTool."
authors = ["Thomas Lehoux, Daniel Hoherd"]

[tool.poetry.dependencies]
python = "^3.9.0"
ansible = "^8.5.0"
pre-commit = "^3.5.0"
pytest = "^7.4.2"
molecule = "^6.0.2"
docker = "^6.1.3"
molecule-docker = "^2.1.0"
pytest-testinfra = "^9.0.0"

[tool.black]
line-length = 132
target-version = ['py39']
include = '\.pyi?$'
exclude = '(venv|.vscode)'

[build-system]
requires = ["poetry>=1.3.2"]
build-backend = "poetry.masonry.api"
